data for gene_presence test case
