+++ test/unit/all_jobs/transform_job/test_dataframes_job.py
@@ -5,14 +5,8 @@ from dagster import RunConfig
 
 from synphage.jobs import (
     PipeConfig,
-    gene_presence,
     transform,
 )
-
-
-@pytest.mark.skip(reason="no way of currently testing this")
-def test_gene_presence():
-    pass
 
 
 @pytest.mark.skip(reason="no way of currently testing this")
